backend/requirements.txt
# requirements.txt
fastapi[all]==0.70.0
sqlalchemy[asyncio]
aiomysql
python-jose
passlib
alembic
